Minor: Minor modifications to the project configurations
Modify the configuration files related to the project - improves
the process being followed to generate executables.

The bash script is still botch-y at best, needs a near-total redo
to be of some actual use, pushing it to the back-log for now.

Update the cmake file with the recent changes being made to the
project - cmake can be used to get the latest version of the project
running directly.