Update: Add play fair cipher
Add a new header file to contain the code for play fair cipher.
This header file introduces a new method to encrypt a string
using play fair cipher mechanism.

Excluding this, the source file contains additional internal
methods that are used to encrypt text using play fair cipher
mechanism, and effectively some other helper methods useful
for debugging the code.

Add a new helper method to generate a copy of a string literal;
excessive use of this header could lead to a memory leaks
and/or StackOver Flow unless the strings being generated by this
method are destroyed.