Update: Add verbose mode for play-fair cipher. Other minor changes
Not really needed, but decided to add a verbose mode to play-fair
as a personal challenege nevertheless. Yet to see if this can be
scaled to other cipher algorithms or not.

Polish the internal code documentation at certain places, minor
other changes to slightly refactor the code base for ease of
debugging (which might be needed in the future).

Slight modifications to the cmake-list to ensure that the project
is built using C-11, not really needed - just a flag to make sure
that only a specific version is being used. In case of a bug, I can
directly proceed to debugging instead of wondering if the cause
could be a deprecated part of the compiler.