BugPatch: Patch a minor bug with the noob-mode. Minor refactor.
In the prior commit introducing noob-mode, a bug managed to
pass through unnoticed (thanks to the lack of test cases).
The bug didn't cause any breaking changes - but, it ensured
that even without the noob mode, the program kept spamming the
console with the steps being performed (missing `if` conditions).
This commit intendes to patch that, users that opt not to go with
the noob-mode, will not be spammed with useless output.

Additionally, since all the cipher algorithms are planned to be
added to a single header file (which will expose only the releavant
stuff to the rest of the program - as an API), this meant that the
source files could not use overlapping names either - a big drawback,
honestly. The lack of C++ like namespaces stings in here -
nevertheless this commit makes minor modifications to the variable
names being used in Play Fair cipher - intending to reduce the common
variable names by adding a cipher-specific prefix to the variables.