Minor: Add a new helper method
Add a new helper method to replace/emulate the functionality of
`fgets` method - the main problem with the former being that it
reads new-line character present at the end of the string, as a
part of the string.

This leads to messy and unpredictable output while printing or
processing the value of the string.

Which lead to the creation of a separate inline helper method
to emulate the functionality of the `fgets` method by reading
from stdin and appending the characters to a string as needed.

Another way to overcome this would have been to create a
modified `printf` method that stripped off trailing new-lines
as needed.