Major: Introduce console-line parameters and interactive mode
This is a major change with regards to the user-experience of this
project.

This commit introduces command-line arguments, as well as interactive
mode to request values from the user during runtime.

Additionally, it is possible to use both these modes together,
parameters can be selectively passed in as command-line arguments
while running project. Any (important) missing parameters detected
by the program during its runtime will be requested from the user
in interactive mode.

Both these modes also involve a heavy usage of the regex logic
implemented earlier, and various other helper methods.

Define a basic structure that will be used to store data after
requesting users for the same, process and parse it appropriately
and use it as needed.

Add header file and source file to the cmake-list file.