Update: Add Railfence cipher. Major bug-fixes.
Introduces major bug-fixes. Major one of them being; the output
always used the verbose mode even if the verbose mode had not
been selected by the user. This was being caused as the value of
verbose mode had not been initialized - logical bug, that caused
statements like `if (verbose)` to go through.

Add railfence cipher to this branch. Involves a minor change of logic
while accepting user-input. In case of railfence, the key is to be
a number (instead of an alphabetical string) - whereas in up till this
point, the data input was not prepared to recieve numeric keys.

Instead of undergoing a major refactor of logic to accodomate for this
change, the railfence source file reveals another method that is to be
used to validate the key that has been passed in. In case the key is
non-numeric, the flow of control will be stopped, and the user will
be asked to re-run with a numeric key.