Major: Large refactor. Add shell-script compiler. Update Makefile
This commit introduces large scale changes to the existing codebase.

Modify the CMakeLists to now target debug and release modes, other
minor changes to adapt it better to the codebase. Including an
attempt to document it better :v

Update the shell script to be able to directly compile a build
using the cmakelist as it exists.

Patch a minor bug in RailFence cipher algorithm, ensuring that the
generated cipher can be deciphered while being scalable. Minor
changes to the excessive (and wasteful) output that was being
dumped in verbose mode - ensuring that only the releavant info is
displayed.